requirements.txt
streamlit==1.43.2
langchain==0.3.22
langchain-google-genai==2.0.10
chromadb==1.0.10
PyMuPDF==1.25.4
python-dotenv==1.0.1
requests==2.32.3
python-docx==1.1.2
google-generativeai==0.8.5
